+++ All.ui/components/colorPicker/colorPicker.js
@@ -9,7 +9,8 @@ const deNone  = '#ffffff' // 当 default 为 false 时 显示
 const app = {
   h: 0, // 0 - 360
   s: 0, // 0 - 100
-  v: 0 // 0 - 100
+  v: 0, // 0 - 100
+  a: 1  // 0 - 1
 } // 全局变量 hsv 的值
 
 Component({
@@ -32,6 +33,10 @@ Component({
     default: {
       type: Boolean,
       value: true
+    },
+    showAlpha: {
+      type: Boolean,
+      value: false
     }
 
   },
@@ -45,13 +50,17 @@ Component({
 
     ouColor: {
       hex: '',
-      rgb: ''
-    }, // rgb与hex 颜色 | 最后输出的颜色
+      rgb: '',
+      rgba: ''
+    }, // rgb, rgba与hex 颜色 | 最后输出的颜色
     bcColor: '', // 颜色面板 背景颜色
     noColor: deNone, // !default || isClear = dom显示 deNone 颜色
-    seColor: deColor, // ouColor.hex 无值时显示 
+    prColor: deColor, //  弹出view 遮罩层颜色 + 复制颜色
+    
+    alphaC: '', // slider标签的背景颜色
 
     hValue: 0, // slider标签 的 value值 取值范围 0-360
+    aValue: 100, // slider标签的 value值 取值范围 0 -100
 
     wSpeed: 0, // 宽度步长
     hSpeed: 0, // 高度步长
@@ -104,18 +113,15 @@ Component({
 
         this.onPredfinedColor()
       }))
-
-
-
     },
 
     // 设置预定颜色
     onPredfinedColor: function () {
-      let { predefined, noColor, seColor } = this.data
+      let { predefined, noColor } = this.data
 
-      // 设置无默认值 + 没有改变颜色 = seColor 颜色
+      // 设置无默认值 + 没有改变颜色 = deColor 颜色
       if ((!this.data.default && predefined === noColor) || !predefined) {
-        predefined = seColor
+        predefined = deColor
       }
 
       let c = predefined
@@ -129,38 +135,58 @@ Component({
         this.setRGB_XY(Color.hex_rgb([h, e, x]))
 
       } else if (c.match(/^[rR][gG][Bb][\(]([\d+,]*?)[\)]$/ig) !== null) {
-        let str = c.split(','),
-                  r = str[0].split('(')[1],
-                  g = str[1],
-                  b = str[2].split(')')[0];
 
-        this.setRGB_XY([r, g, b].map(this.checkRGB))
+        this.setRGB_XY(this.clearRgb(c).map(this.checkRGB))
 
+      } else if (c.match(/^[rR][gG][Bb][aA][\(]([1-9]\d*\.?\d*)|(0\.\d*[1-9]*?)[\)]$/ig) !== null) {
+
+        let arr = this.clearRgba(c)
+        let a = this.checkRGBA(arr.pop())
+        arr.map(this.checkRGB)
+
+        this.setRGB_XY(arr, a)
+       
       } else {
         throw new Error(`The color type is wrong, it should be hex or rgb --- 颜色类型错误， 它应该是 16进制 或 RGB `)
       }
-
     },
 
-    // 校验 rgb的合法值
-    checkRGB: function (n) {
-      n = parseInt(n)
-      if (n > 255) throw new Error(`The value of rgb is wrong, it should be 0-255 --- rgb的值是错误的，范围0-255`)
-      else return n
-    },
-
+    
     // setData x 和 y 的坐标
-    setRGB_XY: function ([r, g, b]) {
+    setRGB_XY: function ([r, g, b], a = app.a) {
       const { h, s, v } = Color.rgb_hsv(r, g, b)
 
       app.h = h
+
+      this.data.ouColor.hex = Color.rgb_hex([r, g, b])
+      
+      this.setAlpha(r, g, b, a)
+
       this.setData({
         hValue: app.h,
+        aValue: this.data.aValue,
         bcColor: Color.hsv_rgb(app.h, svRange, svRange).rgb,
-        'ouColor.hex': Color.rgb_hex([r, g, b]),
+        prColor: this.data.prColor,
+        alphaC: this.data.alphaC,
         x: Math.round(s * this.data.wSpeed),
         y: Math.round((svRange - v) * this.data.hSpeed)
       })
+    },
+
+
+    // 设置透明度颜色
+    setAlpha: function(r, g, b, a) {
+      if (this.data.showAlpha) {
+        let str = `rgba(${r}, ${g}, ${b}`
+        this.data.prColor = `${str}, ${a})`
+        this.data.alphaC = `${str}, 0), ${str}, 1)`
+      } else {
+        this.data.prColor = this.data.ouColor.hex
+        this.data.alphaC = ''
+      }
+
+      if (a == 1) this.data.aValue = 100
+      else this.data.aValue = a * 100
     },
 
 
@@ -173,16 +199,19 @@ Component({
       app.s = Math.round(x / this.data.wSpeed)
       app.v = svRange - Math.round(y / this.data.hSpeed)
 
-      const { rgb, hex } = Color.hsv_rgb(app.h, app.s, app.v)
-      const data = {
-        'ouColor.rgb': rgb
+      const { rgb, hex, rgba } = Color.hsv_rgb(app.h, app.s, app.v, app.a)
+      this.data.ouColor = { rgb, hex, rgba }
+
+      const [r, g, b] = this.clearRgb(rgb)
+      this.setAlpha(r, g, b, app.a)
+
+      // 非 setData 触发 就更新 prColor颜色
+      if (source) {
+        this.setData({
+          prColor: this.data.prColor,
+          alphaC: this.data.alphaC
+        })
       }
-
-      // 非 setData 触发 就更新 hex颜色
-      if (source) data['ouColor.hex'] = hex
-
-      this.setData(data)
-
     },
 
 
@@ -191,47 +220,122 @@ Component({
       const { value } = detail
       app.h = value
 
-      const { rgb, hex } = Color.hsv_rgb(app.h, app.s, app.v)
+      const { rgb, hex, rgba } = Color.hsv_rgb(app.h, app.s, app.v, app.a)
+
+      this.data.ouColor = { rgb, hex, rgba }
+
+      const [r, g, b] = this.clearRgb(rgb)
+      this.setAlpha(r, g, b, app.a)
 
       this.setData({
         bcColor: Color.hsv_rgb(app.h, svRange, svRange).rgb,
-        'ouColor.rgb': rgb,
-        'ouColor.hex': hex
+        prColor: this.data.prColor,
+        alphaC: this.data.alphaC
       })
     },
+
+
+    // 获取 移动 A 的坐标 | 透明度
+    onTouchMoveA: function({ detail }) {
+      app.a = Number((detail.value * 0.01).toFixed(1))
+      
+      const rgb = this.clearRgb(this.data.ouColor.rgb)
+      this.data.ouColor.rgba = `rgba(${rgb.join()}, ${app.a})`
+  
+      this.setData({
+        prColor: this.data.ouColor.rgba
+      })
+    },
+    
 
     // 复制 hex 的值
     onCopyHex: function () {
-      const { ouColor, seColor } = this.data
+      const { prColor } = this.data
       wx.setClipboardData({
-        data: ouColor.hex || seColor
+        data: prColor
       })
     },
 
+
     // 清空颜色，传出回调
     onClear: function () {
-      let { ouColor, isClear, predefined, seColor } = this.data
+      let { ouColor, isClear, predefined, prColor, alphaC } = this.data
 
-      ouColor = { hex: '', rgb: '' }
+      ouColor = { hex: '', rgb: '', rgba: '' }
+      prColor = ''
+      alphaC  = ''
       isClear = true
-      predefined = seColor
-      this.setData({ predefined, isClear, ouColor })
+      app.a = 1
+      predefined = deColor
+      this.setData({ predefined, isClear, prColor, alphaC })
       this.triggerEvent('change', ouColor)
       this.showColorPicker()
     },
+
 
     // 确认颜色，传出回调
     onConfirm: function () {
       /*
             bindchange: 当颜色值改变时触发
       */
-      let { ouColor, isClear, predefined } = this.data
+      let { ouColor, isClear, predefined, prColor } = this.data
 
       isClear = false
-      predefined = ouColor.hex
+      predefined = prColor
       this.setData({ predefined, isClear })
       this.triggerEvent('change', ouColor)
       this.showColorPicker()
+    },
+
+    // 校验 rgb的合法值
+    checkRGB: function (n) {
+      n = Number(n)
+      if (n <= 255 && n >= 0) {
+        return n
+      } else {
+        throw new Error(`The value of rgb is wrong, it should be 0-255 --- rgb的值是错误的，范围0-255`)
+      }
+    },
+
+    //校验rgba的合法值
+    checkRGBA: function(n) {
+      n = Number(n)
+      if (n <= 1 && n >= 0) {
+        return n
+      } else {
+        throw new Error(`The value of rgba is wrong, it should be 0-1 --- rgba的值是错误的，范围0-1`)
+      }
+    },
+
+    /**
+     * 清洗 rgb
+     * @param {string} c rgb(255, 255, 255) 
+     */
+    clearRgb: function(c) {
+      if (c.split(',').length !== 3) {
+        throw new Error(`Please enter the correct RGB color --- 请输入正确的 rgb 颜色`)
+      }
+      let str = c.split(','),
+                r = str[0].split('(')[1],
+                g = str[1],
+                b = str[2].split(')')[0];
+      return [r, g, b]
+    },
+
+    /**
+     * 清洗 rgba
+     * @param {string} c rgb(255, 255, 255, 1) 
+     */
+    clearRgba: function(c) {
+      if (c.split(',').length !== 4) {
+        throw new Error(`Please enter the correct RGBA color --- 请输入正确的 rgba 颜色`)
+      }
+      let str = c.split(','),
+                r = str[0].split('(')[1],
+                g = str[1],
+                b = str[2],
+                a = str[3].split(')')[0]
+      return [r, g, b, a]
     }
   }
 })